% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnoliaBuild.R
\name{magnoliaBuild}
\alias{magnoliaBuild}
\title{Build a Magnolia model}
\usage{
magnoliaBuild(code, filename = NA)
}
\arguments{
\item{code}{Source CSL code (string) for the model}

\item{filename}{Name of a file containing the model source code; if a value is supplied for this argument, the @code argument is ignored.}
}
\value{
An executable Magnolia model object
}
\description{
This function builds an executable Magnolia model object
from the passed CSL source code.
}
