% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnoliaResults.R
\name{magnoliaResults}
\alias{magnoliaResults}
\title{Return a Data Frame Containing Simulation Results}
\usage{
magnoliaResults(mdl)
}
\arguments{
\item{mdl}{A model object created using the magnoliaBuild() function}
}
\value{
An Data Frame containing simulation results
}
\description{
This function creates a data frame containing the time histories
of any model outputs that were included on the prepare list (i.e.,
logged) as of the last simulation run.  The magnoliaResults function
is provided as an alternative to collecting time histories one
at a time using the history() method of the model object.
}
\details{
Note that the model needs to have had ouputs flagged for logging
using the prepare() method, and the model needs to have been executed
for any results to be available.
}
